Refactor the Application class hierarchy to better separate Public/ConfidentialClientApplication and ManagedIdentityApplication
