+++ msal4j-sdk/src/main/java/com/microsoft/aad/msal4j/MsalRequest.java
@@ -15,28 +15,28 @@ abstract class MsalRequest {
 
     AbstractMsalAuthorizationGrant msalAuthorizationGrant;
 
-    private final AbstractClientApplicationBase application;
+    private final AbstractApplicationBase application;
 
     private final RequestContext requestContext;
 
     @Getter(value = AccessLevel.PACKAGE, lazy = true)
     private final HttpHeaders headers = new HttpHeaders(requestContext);
 
-    MsalRequest(AbstractClientApplicationBase clientApplicationBase, AbstractMsalAuthorizationGrant abstractMsalAuthorizationGrant, RequestContext requestContext) {
+    MsalRequest(AbstractApplicationBase clientApplicationBase, AbstractMsalAuthorizationGrant abstractMsalAuthorizationGrant, RequestContext requestContext) {
         this.application = clientApplicationBase;
         this.msalAuthorizationGrant = abstractMsalAuthorizationGrant;
         this.requestContext = requestContext;
 
         CurrentRequest currentRequest = new CurrentRequest(requestContext.publicApi());
-        application.getServiceBundle().getServerSideTelemetry().setCurrentRequest(currentRequest);
+        application.serviceBundle().getServerSideTelemetry().setCurrentRequest(currentRequest);
     }
 
-    MsalRequest(AbstractClientApplicationBase clientApplicationBase, RequestContext requestContext) {
+    MsalRequest(AbstractApplicationBase clientApplicationBase, RequestContext requestContext) {
         this.application = clientApplicationBase;
         this.requestContext = requestContext;
 
         CurrentRequest currentRequest = new CurrentRequest(requestContext.publicApi());
-        application.getServiceBundle().getServerSideTelemetry().setCurrentRequest(currentRequest);
+        application.serviceBundle().getServerSideTelemetry().setCurrentRequest(currentRequest);
     }
 
     MsalRequest() {
