+++ msal4j-sdk/src/main/java/com/microsoft/aad/msal4j/RemoveAccountRunnable.java
@@ -9,7 +9,7 @@ import java.util.concurrent.CompletionException;
 class RemoveAccountRunnable implements Runnable {
 
     private RequestContext requestContext;
-    private AbstractClientApplicationBase clientApplication;
+    private AbstractApplicationBase clientApplication;
     IAccount account;
 
     RemoveAccountRunnable(MsalRequest msalRequest, IAccount account) {
