+++ msal4j-sdk/src/main/java/com/microsoft/aad/msal4j/TokenRequestExecutor.java
@@ -55,8 +55,9 @@ class TokenRequestExecutor {
         oauthHttpRequest.setContentType(HTTPContentType.ApplicationURLEncoded.contentType);
 
         final Map<String, List<String>> params = new HashMap<>(msalRequest.msalAuthorizationGrant().toParameters());
-        if (msalRequest.application().clientCapabilities() != null) {
-            params.put("claims", Collections.singletonList(msalRequest.application().clientCapabilities()));
+        if (msalRequest.application() instanceof AbstractClientApplicationBase
+                && ((AbstractClientApplicationBase) msalRequest.application()).clientCapabilities() != null) {
+            params.put("claims", Collections.singletonList(((AbstractClientApplicationBase) msalRequest.application()).clientCapabilities()));
         }
 
         if (msalRequest.msalAuthorizationGrant.getClaims() != null) {
@@ -78,7 +79,8 @@ class TokenRequestExecutor {
 
         oauthHttpRequest.setQuery(URLUtils.serializeParameters(params));
       
-        if (msalRequest.application().clientAuthentication() != null) {
+        if (msalRequest.application() instanceof AbstractClientApplicationBase
+                && ((AbstractClientApplicationBase) msalRequest.application()).clientAuthentication() != null) {
 
             Map<String, List<String>> queryParameters = oauthHttpRequest.getQueryParameters();
             String clientID = msalRequest.application().clientId();
@@ -92,7 +94,7 @@ class TokenRequestExecutor {
                         .createClientAuthFromClientAssertion((ClientAssertion) ((ClientCredentialRequest) msalRequest).parameters.clientCredential())
                         .applyTo(oauthHttpRequest);
             } else {
-                msalRequest.application().clientAuthentication().applyTo(oauthHttpRequest);
+                ((AbstractClientApplicationBase) msalRequest.application()).clientAuthentication().applyTo(oauthHttpRequest);
             }
         }
         return oauthHttpRequest;
